implemented basic routing
